feat: Show info bar in Word Masters game
The code changes in this commit modify the Word Masters game by showing the info bar. The info bar was previously hidden, but now it is set to be visible by adding the "show" class to the info-bar element in the HTML and CSS files.

This change was made to enhance the user experience and provide important information to the players during the game.